requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.10.2
MarkupSafe==3.0.3
Pygments==2.17.2
pymdown-extensions==10.21
python-frontmatter==1.1.0
pytz==2026.1.post1
PyYAML==6.0.3
Werkzeug==3.1.6
